+++ example/lib/main.dart
@@ -8,6 +8,7 @@ import 'package:example/button_section.dart';
 import 'package:example/color_section.dart';
 import 'package:example/icon_shape_section.dart';
 import 'package:example/input_field_section.dart';
+import 'package:example/section_container.dart';
 import 'package:example/spinner_section.dart';
 import 'package:example/typography_section.dart';
 import 'package:flowbite_ui/flowbite_ui.dart';
@@ -32,6 +33,10 @@ class MainApp extends StatelessWidget {
           child: Column(
             spacing: 24.0,
             children: [
+              SectionContainer(
+                fileName: 'accordion',
+                child: PreviewFlowbiteAccordion(),
+              ),
               AlertSection(),
               AvatarSection(),
               AvatarGroupLabelSection(),